requirements.txt
requests>=2.28.0
pandas>=1.5.0
numpy>=1.21.0
python-binance>=1.0.16
python-dotenv>=0.19.0
mplfinance>=0.12.9b7
seaborn>=0.11.0
matplotlib>=3.5.0 